Updates to strategy.
Adding new function to find closest distance between point and line.
Can also use an existing function to get closest point and then find
distance but this is easier to read.
In Game, adding check to see if we are close to any of the walls or our
opponent.
If so, we back off.