Updating Communicator2(). 	 - Seems to work but needs testing. 	 - Moataz, you can implement your work on this class.
Updating Controllers to include setWheelSpeeds() method. This should ensure that we shouldn't go over our max speed that's set in the Roboto class.

Tried to fix touch sensors so they should stop after 150ms (whether we are running strategy or not).

Added PenaltyKickStraight() - This allows us to kick in the same direction we are placed in intead of randomly rotating by 27 degrees.
      - This needs fully implemented (e.g. as strategy).