Second attempt at integrating PID controller
Forgot to add my first attempt (WallE.java). That worked in most cases
but I think as it was threaded it crashed when receiving too many
instructions.
This controller is simple. It's the same as before, but if we're going
forward, we update the wheel speeds and then continue as normal!
Testing this weekend.